build: keep FileAccess runtime surface in runtime lib; ast_validate to compiler
The file-access runtime surface must live in the runtime lib so C++ embedders
that link only the runtime lib can construct it to run scripts. Move it back
out of the compiler lib:

- file_access.cpp (FileAccess base impl + vtable) and fs_file_info.cpp
  (FsFileAccess, addFsRoot) -> runtime (SIMULATE_SRC); module_file_access.cpp
  (ModuleFileAccess) -> runtime (BUILTIN_SRC).
- De-virtualize FileAccess::serialize / ModuleFileAccess::serialize, mirroring
  the FileInfo treatment: a serializeKind tag replaces the vtable slot so the
  runtime types do not reference the compiler-only serializer; the body lives
  as serializeFileAccessBody in compiler-lib ast_serialize.cpp, dispatched by
  the tag. FileInfo/TextFileInfo get the same serializeKind de-virtualization.
- get_file_access (pak compile, ast_parse.cpp), the AstSerializer, and the
  daslang FileAccess bindings stay compiler-side.

Logic-op function annotations (LogicOp*Annotation, newLogicAnnotation) move
from module_builtin_runtime.cpp to module_builtin_ast.cpp (compiler lib) - they
are pure compile-time.

ast_validate.cpp moves from AST_SRC (both libs) to the compiler lib; only
Program::validateAst is exported and its sole caller is compiler-side
ast_parse.cpp.

Verified: full link (runtime + compiler Dyn libs), tests/language 1060/1060,
serializer round-trip over ./tests 8899/8899 pass.
