+++ src/builtin/module_file_access.cpp
@@ -1,6 +1,5 @@
 #include "daScript/misc/platform.h"
 #include "daScript/ast/ast.h"
-#include "daScript/ast/ast_serializer.h"
 #include "daScript/simulate/debug_info.h"
 
 #if !defined(DAS_NO_FILEIO)
@@ -12,6 +11,7 @@ das::Context * get_context ( int stackSize=0 );
 namespace das {
 
     ModuleFileAccess::ModuleFileAccess() {
+        serializeKind = 1;
     }
 
     ModuleFileAccess::~ModuleFileAccess() {
@@ -19,6 +19,7 @@ namespace das {
     }
 
     ModuleFileAccess::ModuleFileAccess ( const string & pak, const ProgramPtr & program ) {
+        serializeKind = 1;
         TextWriter tout;
         if ( program ) {
             if ( program->failed() ) {
@@ -215,23 +216,6 @@ namespace das {
         auto exc = context->getException(); exc;
         DAS_ASSERTF(!exc, "exception failed in `pod_in_scope_allowed`: %s", exc);
         return cast<bool>::to(res);
-    }
-
-
-    void FileAccess::serialize ( AstSerializer & ser ) {
-        if ( ser.writing ) {
-            uint8_t tag = 0;
-            ser << tag;
-        }
-        ser << files;
-    }
-
-    void ModuleFileAccess::serialize ( AstSerializer & ser ) {
-        if ( ser.writing ) {
-            uint8_t tag = 1;
-            ser << tag;
-        }
-        ser << files;
     }
 
 }
